LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/elements-in-the-dom/unknown-element-expected.txt


PASS <applet> is an HTMLUnknownElement
PASS <attachment> is an HTMLUnknownElement
PASS <layer> is an HTMLUnknownElement
PASS <nolayer> is an HTMLUnknownElement
PASS <xxx> is an HTMLUnknownElement
